requirements.txt
setuptools==39.0.1
numpy==1.13.3
opencv_python==3.4.2.17
Flask==1.0.2
torchvision==0.2.1
torch==0.4.1
scipy==1.1.0
matplotlib==2.2.3
Pillow==5.3.0
